// apps/funnel/utils/enums/finby-result-codes.ts
/**
 * finby result codes (from documentation)
 * Based on https://doc.finby.eu/#overview-intro
 */
export enum FinbyResultCodes {
    SUCCESS = 0,
    PENDING = 1,
    PROCESSING = 4,
    AUTHORIZED_ONLY = 5,
    INVALID_REQUEST = 1001,
    UNKNOWN_ACCOUNT = 1002,
    MERCHANT_ACCOUNT_DISABLED = 1003,
    INVALID_SIGNATURE = 1004,
    USER_CANCEL = 1005,
    INVALID_AUTHENTICATION = 1006,
    INSUFFICIENT_BALANCE = 1007,
    SERVICE_NOT_ALLOWED = 1008,
    PROCESSING_ID_USED = 1009,
    TRANSACTION_NOT_FOUND = 1010,
    UNSUPPORTED_TRANSACTION = 1011,
    REJECTED_TRANSACTION = 1014,
    DECLINED_BY_RISK = 1015,
    GENERAL_ERROR = 1100,
    UNSUPPORTED_CURRENCY_CONVERSION = 1101,
}